move flagString utility into shared space
